feat(Generator): Add * operator (amplitude scaling) and mock processing
Ex: `sine * 0.5` will scale the amplitude of the sine wave.

Setting mock processing for a generator will cause the RootNode to process the generator
without using the processed sample. This is useful when needing to
automate the process() function of the generator via Engine's process
but the output of the generator is not needed.