feat(Logic and Polynomial): Update processors to safely synchronize with external nodes
NodeState Final feature Pass
When using internal nodes for either Logic or Polynomial, the processors
now correctly update to the new internal node on the next process call.
This ensures single process call, appropriate reset and correct signal
flow.
Its now also allowed to use external nodes that are connected to other nodes
in the processing graph. This allows for complex signal flow where the
processor's input is determined by the state of the external node.
Here the synchronization is guaranteed but the flow is dependent on
outside connections to the node,