refactor(KakshyaUtils): Split to multiple files
KakshyaUtils was getting much larger and was still incomplete.
This refactor also adds new functions to help with managing regions and data.
It splits based on functionality and moved related functions to their own files.
Several redundant usages from Processors and containers are replaced or
redirected to the new files.