feat(variant vector WIP): Utils and abstract container refactor
DO NOT USE:
This is the first pass of many that enables containers to use vector of
variants as data instead of single variant. All the non concrete
containers are prepared with relevant updates to return types, region
access, channel tracking, position tracking and memory layout.

Kakshya Utils functions are upgraded, refactored or discarded based on
multi dimensional datavariant support