feat(OperationHelper): MultiVariant support
This is a complete rewrite of OperationHelpers to use type_traits,
support vector of variants, combine several methods into single branch
and removal of legacy features.
RegionUtils and Containers now have dedicated but simplistic methods to
extract segments and region groups. For comprehensive extraction it is
still recommended to use RegionProcessors