refactor(API): Update initializer and context appliers
Context appliers for Creater/Proxy require the engine to supply
registartion functions that are needed to be called at Init.
The current model works for regular C++ workflow but does not for
Interpreter workflow.
Upadting to call the appliers on including MayaFlux header with
MAYASIMPLE declaration.
The applier methods are now static members instead of local static, to
expose them cross context