docs(Tutorial): Sculpting Data II: Processing Expressions
This tutorial explores buffer processors, chains and related semantics
introduced in Part I, and expands coverage into Input buffers and buffer
clone, supply, and mix semantics.

But more importantly, this tutorial introduces the concepts of using
mathematical expressions as creative material to either sculpt or to
create audio signals.
It also introduces logic as more than branching structures, but as a way to
make creative or artistic decisions in the context of audio signal processing.