feat(ToeknDistributor): Pass I
New class: BufferTokenDistributor. This handles token setup
when requested buffer tokens do not align with the expectations of the
root buffer for that domain.
For instance, when audio is requested as AUDIO_PARALLEL, it requires
processing at sample rate but cannot be done on the cpu. The token
distributor is meant to parse this situation, create an appropriate
transfer processor and mark the buffer as needing internal processing.

This is not integrated into buffer management yet as this is a complex
feature with several gotchas to consider. Entirely possible it is only
stubs for 0.1.0 release