BREAKING CHANGE: vega.read() -> vega.read_audio()
This is necessary to allow for read_* to support images, videos, generic
files in the upcoming PRs.
Attempts were made to use variants, builder patterns, and other techniques, but
ultimately the function signature had to change to accommodate the new design.

All documentation and examples have been updated to reflect this change.
Weave templates will be updated in that repo subsequently.