feat(Rendering): Allow specialized buffers to own and deploy RenderingProcessor
Instead of the verbose mechanism of creating a new rendering processor
with the same 4-8 default parameters, registering and attaching it to
TextureBuffer, GeometryBuffer or NetworkGeometryBuffer, we now allow these
buffers to own a RenderingProcessor directly. This simplifies the code and
makes it easier to create specialized buffers with custom rendering logic.