fix(AudioBuffer): Add setup_processor redirect
To avoid calling shared_from_this() in constructor
of audio buffer when attaching processor, the parent class now
provided a setup_processor method that is called after construction by
default in BufferManager when adding/registering a new buffer.

All child classes of AudioBuffer that require specific processor setup
should override this method instead of doing it in the constructor.