fix(Shaders): Reorder pipeline config and descriptor binding
In the last refactor of shader processor, we moved from all inherited
classes declaring their own processing_function to a single parent and
multiple execute_shader overrides. but this broke many special orderings
required by various children. This is a fix for render processor to
still maintain same flow in the new system.