Deps(Vulkan): Replace spirv-reflect with spirv-cross Phase 1
Currently the only way to use spirv-reflect is via fetch content as no distros or package managers bundle/provide-builds for it.
fetch content is not so legally far from vendoring, also disallowed by many automated buiild systems.

This is the first of several commits that will systematically replace spirv-reflect with spirv-cross as all package mangers have it, windows bundles it via lunarg vulkan sdk, and the package provides all features mayaflux for reflection