feat(Tutorial): VisualMateriality, graphics tutorials pass I
This is the first commit, to be followed by refinements as necessary,
towards the part I of visual materiality tutorials. This is still part
of the sculpting data series, it just kicks off a series of graphics
tutorials from simple geometry, to shaders to textures and materials.

The commit also includes some quick fixes to minor issues revealed
while working on the tutorials, including a bug in the particle
network where the buffer was not being properly resized when the
number of particles changed.